HOTFIX: Ban 의 챔피언 타입 int -> Long

+++ src/main/java/com/clonelol/summoner/api/summonerapi/dto/property/infoProperty/teamProperty/Ban.java
@@ -1,6 +1,6 @@
 package com.clonelol.summoner.api.summonerapi.dto.property.infoProperty.teamProperty;
 
 public class Ban {
-    private int championId;
+    private Long championId;
     private int pickTurn;
 }
